feedlist seperation complete
