+++ src/navigation/index.tsx
@@ -2,7 +2,7 @@ import { useNavigation } from '@react-navigation/native';
 import React, { useEffect, useState } from 'react';
 import { Pressable, ScrollView, View, Text } from 'react-native';
 import { useTheme } from 'react-native-paper';
-import { getAccessToken } from '../utils/hooks/asyncStorage/Login';
+import { getAccessToken, getUserId } from '../utils/hooks/asyncStorage/Login';
 
 export function RootScreen() {
   const navigation = useNavigation();
@@ -24,6 +24,12 @@ export function RootScreen() {
         onPress={() => getAccessToken().then((res) => console.log(res))}
       >
         <Text style={theme.fonts.text.display}>Get Token</Text>
+      </Pressable>
+      <Pressable
+        style={{ borderWidth: 1, backgroundColor: 'grey', width: 300, height: 50 }}
+        onPress={() => getUserId().then(console.log)}
+      >
+        <Text style={theme.fonts.text.display}>Get USER ID</Text>
       </Pressable>
 
       <Text style={theme.fonts.text.body1.regular}>{accessToken}</Text>
