+++ src/screens/Feed/FeedList.tsx
@@ -1,21 +1,52 @@
-import { StyleSheet, Text, View } from 'react-native';
-import React, { Fragment } from 'react';
-import { InfiniteData, UseMutateFunction } from '@tanstack/react-query';
+import { RefreshControl, StyleSheet, Text, View } from 'react-native';
+import React, { Fragment, useState } from 'react';
+import { InfiniteData, QueryClient, UseMutateFunction } from '@tanstack/react-query';
 import { InsightData } from '../../types/Feed/Feedinsights';
-import { InView } from 'react-native-intersection-observer';
+import { IOScrollView, InView } from 'react-native-intersection-observer';
 import FeedItem from './FeedItem';
+import { FeedQueryKeys } from '../../utils/api/FeedAPI';
+import { UserSpecificChallengeQueryKeys } from '../../utils/api/UserSpecificChallenge';
+import DividerBar from '../../components/bars/DividerBar';
 
 interface FeedListProps {
   feedList: InfiniteData<InsightData[] | undefined> | undefined;
   fetchNextPage: () => void;
   touchBookMark: UseMutateFunction<void, unknown, number, unknown>;
-
   bookMarkIsLoading: boolean;
+  // UpperComponent is not required, but used in FeedScreen.tsx for to display the user's current streak
+  UpperComponent?: React.ReactNode;
+  feedListQueryClient: QueryClient;
+  feedListIsLoading: boolean;
 }
 
-const FeedList = ({ feedList, fetchNextPage, touchBookMark, bookMarkIsLoading }: FeedListProps) => {
+const FeedList = ({
+  UpperComponent,
+  feedList,
+  fetchNextPage,
+  touchBookMark,
+  bookMarkIsLoading,
+  feedListQueryClient,
+  feedListIsLoading,
+}: FeedListProps) => {
+  const [pageRefreshing, setPageRefreshing] = useState(false);
+
+  const onRefresh = () => {
+    setPageRefreshing(true);
+    feedListQueryClient.invalidateQueries(FeedQueryKeys.getFeed());
+    feedListQueryClient
+      .invalidateQueries(UserSpecificChallengeQueryKeys.getUserSpecificChallenge())
+      .then(() => setPageRefreshing(false));
+  };
+
   return (
-    <View>
+    <IOScrollView
+      refreshControl={<RefreshControl refreshing={pageRefreshing} onRefresh={onRefresh} />}
+      contentContainerStyle={styles.feedCtn}
+    >
+      {/* UpperComponent will render undefined, which does not affect this component.  */}
+      {UpperComponent}
+      <DividerBar style={styles.divider} />
+
       {feedList?.pages.map((group, i) => {
         return (
           <Fragment key={i}>
@@ -46,10 +77,23 @@ const FeedList = ({ feedList, fetchNextPage, touchBookMark, bookMarkIsLoading }:
           </Fragment>
         );
       })}
-    </View>
+    </IOScrollView>
   );
 };
 
 export default FeedList;
 
-const styles = StyleSheet.create({});
+const styles = StyleSheet.create({
+  feedCtn: {
+    padding: 16.5,
+  },
+  divider: {
+    backgroundColor: '#f8f8f4',
+    borderBottomColor: '#f8f8f4',
+    marginBottom: 24,
+    height: 12,
+    width: '150%',
+    marginLeft: 0,
+    left: -50,
+  },
+});
